Adding/correcting read/write operations

+++ libc/getdents.c
@@ -1,5 +1,7 @@
-#include<libc.h>
+#include <libc.h>
+#include <unistd.h>
+#include "sys/defs.h"
 
-int getdents(int fd, char* dirp, int count) {
+int getdents(int fd, char* dirp, size_t count) {
         return syscall_ici(__NR_getdents, fd, dirp, count);
 }
